Fix location of the functional tests ('org.scoverage' -> 'org/scoverage') and separate the cross-version tests to a separate source set
