+++ src/crossScalaVersionTest/java/org/scoverage/Scala212Test.java
@@ -1,4 +1,4 @@
-package org.scoverage.crossversion;
+package org.scoverage;
 
 public class Scala212Test extends ScalaVersionTest {
     public Scala212Test() {
