+++ src/crossScalaVersionTest/resources/projects/scala-multi-module-cross-version/2_11/build.gradle
