+++ src/crossScalaVersionTest/resources/projects/scala-multi-module-cross-version/2_12/build.gradle
