+++ src/crossScalaVersionTest/resources/projects/scala-multi-module-cross-version/2_13/build.gradle
