+++ src/crossScalaVersionTest/resources/projects/scala-multi-module-cross-version/build.gradle
