+++ src/functionalTest/java/org/scoverage/CompositeBuildTest.java
