+++ src/functionalTest/java/org/scoverage/DetectScalaLibraryTest.java
