+++ src/functionalTest/java/org/scoverage/MultiModulePluginNotConfiguredForScalaTest.java
