+++ src/functionalTest/java/org/scoverage/MultipleCheckTasksTest.java
