+++ src/functionalTest/java/org/scoverage/ScalaJavaMultiModuleTest.java
