+++ src/functionalTest/java/org/scoverage/ScalaMultiModuleTest.java
