+++ src/functionalTest/java/org/scoverage/ScalaMultiModuleWithMultipleTestTasksTest.java
