+++ src/functionalTest/java/org/scoverage/ScalaMultiModuleWithPartialScoverageUseTest.java
