+++ src/functionalTest/java/org/scoverage/ScalaSingleModuleTest.java
