+++ src/functionalTest/java/org/scoverage/ScalaSingleModuleWithDependencyManagerTest.java
