+++ src/functionalTest/java/org/scoverage/ScalaSingleModuleWithMultipleTestTasksTest.java
